requirements.txt
asgiref==3.3.1
Django==3.1.7
pytz==2021.1
sqlparse==0.4.1
pydantic==1.8.1
djangorestframework==3.12.2
flake8==3.9.0
django-stubs==1.7.0
